FILE: renames store.ts to state.ts
