+++ libs/core/util/src/lib/state.ts
